packages/vck-templates/templates/example-agents/blog-generator/requirements.txt
langchain-anthropic>=0.1.0
langgraph>=0.2.0
pydantic>=2.0.0
typing-extensions>=4.0.0
requests>=2.31.0
python-dotenv>=1.0.0
pytest>=7.0.0
